fix(Header...): added logo + icons for mobile + styles corrected
